refactor: split SiloFilterExpressionMapper.kt from AggregatedModel
issue: #205
